Fixed bug making script hang
I fixed a bug that was forcing the script to either hang or to completely kill the PowerShell Session in VSCode. I removed any variables that were added in the code but never declared called $PublishSettings.

The script runs properly now and no longer kills the PowerShell session or hangs.