Created Windows10DebloaterGUI
This is my test file that I have been working with. I am currently creating a GUI for all of the code in this Repository and am working to make it function appropriately. It isn't completely functioning as intended, so before you use it on machines you care about, please make sure you understand that this is in development and I cannot guarantee that things won't break. Also make sure to make a backup.