Fixed Writing to console and Add_Click actions
The application will now write to the host console properly and give updates as it does its work instead of just hanging and acting as if it is frozen. The $RemoveAllBloatware Add_Click now works properly as well as $RemoveBlacklist and the associated Add_Click. 

The console will put a space in between lines when each Add_Click action is finished for better reading and font visibility.