New check for OneDrive Files existence and folder existence
I added code to check for the OneDrive folder and its files before removal. If it finds that the folder exists, it will then check to see if there are files inside of the folder. If there are no files then OneDrive is uninstalled.

If there are files, then it creates a new folder called OneDriveBackupFiles on the user's desktop, move all of the files into that folder, and then removes OneDrive. 

This is a safety measure after a user had their OneDrive files (unfortunately) deleted from their machine.