Fix a syntax error in $WhitelistedApps which prevents Microsoft.MicrosoftStickyNotes and Microsoft.VP9VideoExtensions to be interpreted as whitelisted apps.
The backquote cannot be used inside the string and is interpreted as a character of the regular expression. Because of this two whitelisted apps are not recognized as such. I suppose with this change, the FixWhitelistedApps reinstallation that occurs at the end of the script could be removed as it's no longer needed.